热门手机与连接点业务代码开发

+++ TelecommunicationsBigData-BackEnd/src/main/java/com/bigdata/model/entity/SignalCoverage/TypicalSignalStrengthStatistics.java
@@ -4,7 +4,7 @@ import lombok.Data;
 
 @Data
 public class TypicalSignalStrengthStatistics {
-    private long rssi;
+    private double rssi;
     private String landmark;
     private String networkType;
     private String networkName;
